use np.float64 everywhere to be consistent

+++ pyflux/gas/gasx.py
@@ -62,8 +62,8 @@ class GASX(tsm.TSM):
         self.y, self.X = dmatrices(formula, data)
         self.y_name = self.y.design_info.describe()
         self.X_names = self.X.design_info.describe().split(" + ")
-        self.y = self.y.astype(np.float) 
-        self.X = self.X.astype(np.float) 
+        self.y = self.y.astype(np.float64) 
+        self.X = self.X.astype(np.float64) 
         self.param_no = self.X.shape[1]
         self.data_name = self.y_name
         self.y = np.array([self.y]).ravel()
